requirements.txt
Flask==2.3.2
pytesseract==5.0.0
Pillow==9.3.0
opencv-python==4.7.0.72